00 MANUALES/OLD/requirements-OLd.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
pymysql
cryptography
